Add tests for DurationFieldForm
